api/requirements.txt
black==25.1.0
celery==5.5.2
Django==5.1.6
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-lifecycle==1.2.4
django-model-utils==5.0.0
django-phonenumber-field==8.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
drf-nested-routers==0.94.1
numpy==2.2.5
osrm-py==0.11.3
pandas==2.2.3
phonenumbers==8.13.55
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pylint==3.3.4
pylint-django==2.6.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.0.0
requests==2.32.3
shapely==2.1.0
xlrd==2.0.1
xlwt==1.3.0
